moved database to differnt file, fixed spelling mistake leaved -> left #anton
