+++ server/server.go
@@ -5,7 +5,10 @@ import (
 	pb "example/Mini_Project_2_Chitty-Chat/chat"
 	"log"
 	"net"
-	"sync"
+
+	//"sync"
+
+	cd "example/Mini_Project_2_Chitty-Chat/server/database"
 
 	"google.golang.org/grpc"
 )
@@ -13,6 +16,8 @@ import (
 const (
 	port = ":8008"
 )
+
+var chatData = cd.NewChatDatabase()
 
 type ChatServiceServer struct {
 	pb.UnimplementedChatServiceServer
@@ -33,7 +38,7 @@ func (s *ChatServiceServer) Disconnect(ctx context.Context, user *pb.User) (*pb.
 	chatData.RemoveUser(user)
 	chatData.InsertMessage(&pb.Message{User: user, Text: user.Username + " has left the chat!"})
 	log.Println("Status: " + user.Username + " has left the chat!")
-	return &pb.Response{Status: "Leaved Successful"}, nil
+	return &pb.Response{Status: "Left Successful"}, nil
 }
 
 func (s *ChatServiceServer) Publish(ctx context.Context, msg *pb.Message) (*pb.Response, error) {
@@ -72,78 +77,4 @@ func main() {
 	if err := s.Serve(lis); err != nil {
 		log.Fatalf("failed to serve: %v", err)
 	}
-}
-
-//here
-//begins
-//something
-//special
-
-type chatDatabase struct {
-	connectedUsers  []*pb.User
-	userToMesageMap map[string][]*pb.Message
-	mu              sync.Mutex
-}
-
-var chatData *chatDatabase = &chatDatabase{
-	connectedUsers:  make([]*pb.User, 0),
-	userToMesageMap: make(map[string][]*pb.Message)}
-
-func (cd *chatDatabase) InsertMessage(msg *pb.Message) {
-	cd.mu.Lock()
-	defer cd.mu.Unlock()
-
-	for _, user := range cd.connectedUsers {
-		/*if msg.User.Username == user.Username { //do not send message to the user who wrote it
-			continue
-		}*/
-
-		cd.userToMesageMap[user.Username] = append(cd.userToMesageMap[user.Username], msg)
-	}
-}
-
-func (cd *chatDatabase) PopMessage(user *pb.User) *pb.Message {
-	cd.mu.Lock()
-	defer cd.mu.Unlock()
-
-	if len(cd.userToMesageMap[user.Username]) < 1 {
-		return nil
-	}
-
-	msg := cd.userToMesageMap[user.Username][0]
-	log.Println("Status: Accesing message: " + msg.Text + " - for user: " + user.Username)
-
-	cd.userToMesageMap[user.Username] = cd.userToMesageMap[user.Username][1:]
-	return msg
-}
-
-func (cd *chatDatabase) AddUser(user *pb.User) bool {
-	cd.mu.Lock()
-	defer cd.mu.Unlock()
-
-	for i := range cd.connectedUsers {
-		if cd.connectedUsers[i].Username == user.Username {
-			return false
-		}
-	}
-
-	cd.connectedUsers = append(cd.connectedUsers, user)
-	cd.userToMesageMap[user.Username] = make([]*pb.Message, 0)
-
-	return true
-}
-
-func (cd *chatDatabase) RemoveUser(user *pb.User) {
-	cd.mu.Lock()
-	defer cd.mu.Unlock()
-
-	var newConnectedUsers []*pb.User = make([]*pb.User, 0)
-	for i := range cd.connectedUsers {
-		if cd.connectedUsers[i].Username != user.Username {
-			newConnectedUsers = append(newConnectedUsers, cd.connectedUsers[i])
-		}
-	}
-	cd.connectedUsers = newConnectedUsers
-
-	delete(cd.userToMesageMap, user.Username)
 }
